requirements.txt
appdirs==1.4.4
certifi==2022.9.24
charset-normalizer==2.0.12
colorlog==6.7.0
greenlet==1.1.3
idna==3.4
importlib-metadata==4.8.3
importlib-resources==5.4.0
pkg-resources==0.0.0
psycopg2-binary==2.9.3
pyee==8.2.2
pyppeteer==0.2.6
PyYAML==6.0
requests==2.27.1
SQLAlchemy==1.4.41
tqdm==4.64.1
typing-extensions==4.1.1
urllib3==1.26.12
websockets==9.1
zipp==3.6.0
